pyproject.toml
[tool.poetry]
name = "project3"
version = "0.1.0"
description = ""
authors = ["Zharkov Aleksey <gross3000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
requests = "^2.32.4"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
mypy = "^1.16.1"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
